refactoring/clean code
