+++ src/main/scala/io/blockfrost/sdk/api/MetricsApi.scala
@@ -4,6 +4,7 @@ import io.blockfrost.sdk.ApiClient
 import io.blockfrost.sdk.api.MetricsApi.{EndpointMetric, Metric}
 import io.blockfrost.sdk.common.SttpSupport
 import org.json4s.{Formats, Serialization}
+import sttp.client3.UriContext
 
 trait MetricsApi[F[_], P] extends SttpSupport {
   this: ApiClient[F, P] =>
@@ -11,6 +12,16 @@ trait MetricsApi[F[_], P] extends SttpSupport {
   def getUsageMetrics(implicit formats: Formats, serialization: Serialization): F[ApiResponse[Seq[Metric]]]
 
   def getEndpointUsageMetrics(implicit formats: Formats, serialization: Serialization): F[ApiResponse[Seq[EndpointMetric]]]
+}
+
+trait MetricsApiImpl[F[_], P] extends MetricsApi[F, P] {
+  this: ApiClient[F, P] =>
+
+  def getUsageMetrics(implicit formats: Formats, serialization: Serialization): F[ApiResponse[Seq[Metric]]] =
+    get(uri"$host/metrics")
+
+  def getEndpointUsageMetrics(implicit formats: Formats, serialization: Serialization): F[ApiResponse[Seq[EndpointMetric]]] =
+    get(uri"$host/metrics/endpoints")
 }
 
 object MetricsApi {
